Hydraulic cylinder tech drawing
I was able with some difficulties to make a drawing of hydraylic cylinder with techdraw workbench but I was able, finally. I will try to make a demo example of what I have done so far. See the commit details, because there are several steps involved:
1. Adding a drawing to the CAD file using techdraw workbench.
1.2. Adding references and keywords for other users.
2. Exporting to pdf; exporting to svg.
3. Staging all the files related to this commit (CAD model, pdf, svg).
4. Making the pull request.

AS an interesting remark is good to point that we could have also different techdrawings, of alternative cylinders, and they can be added, or appended as new drawings pages.